resources 추가
