+++ src/main/java/com/sc/secondHandMarket/Product.java
@@ -1,10 +1,13 @@
-// src/main/java/com/sc/secondHandMarket/Product.java
 package com.sc.secondHandMarket;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 
 @Entity
 public class Product {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -12,13 +15,27 @@ public class Product {
     private String name;
     private int price;
 
-    // Getter, Setter
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
+    // 생성자
+    public Product() {}
 
-    public String getName() { return name; }
-    public void setName(String name) { this.name = name; }
+    // Getter / Setter
+    public Long getId() {
+        return id;
+    }
 
-    public int getPrice() { return price; }
-    public void setPrice(int price) { this.price = price; }
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+
+    public void setPrice(int price) {
+        this.price = price;
+    }
 }
