+++ src/main/java/com/sc/secondHandMarket/ProductRepository.java
@@ -1,24 +1,6 @@
-@SpringBootApplication
-public class SecondHandMarketApplication implements CommandLineRunner {
+package com.sc.secondHandMarket;
 
-    private final ProductRepository productRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-    public SecondHandMarketApplication(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    public static void main(String[] args) {
-        SpringApplication.run(SecondHandMarketApplication.class, args);
-    }
-
-    @Override
-    public void run(String... args) {
-        Product product = new Product();
-        product.setName("나이키 에어포스 1");
-        product.setPrice(130000);
-        productRepository.save(product);
-
-        productRepository.findAll().forEach(p ->
-                System.out.println("📦 상품명: " + p.getName() + ", 가격: " + p.getPrice()));
-    }
+public interface ProductRepository extends JpaRepository<Product, Long> {
 }
